+++ server/src/migrations/2020-07-07-201052_add_missing_inv_loadout/down.sql
@@ -1,1 +1,0 @@
--- Nothing to undo since up.sql only creates missing inventory/loadout records