+++ server/src/migrations/2020-07-10-162552_dash_melee_energy_cost_fix/down.sql
@@ -1,1 +1,0 @@
--- Nothing to undo since up.sql only fixes corrupt JSON in loadouts