+++ server/src/migrations/2020-07-16-044718_migrate_armour_stats/down.sql
@@ -1,1 +1,0 @@
--- No down action for this migration