+++ server/src/migrations/2020-12-13-172324_skills/down.sql
@@ -1,1 +1,0 @@
--- What's a down migration?