+++ server/src/persistence/character_loader.rs
@@ -1,7 +1,7 @@
 use crate::persistence::{
     character::{create_character, delete_character, load_character_data, load_character_list},
-    error::Error,
-    establish_connection, PersistedComponents,
+    error::PersistenceError,
+    establish_connection, DatabaseSettings, PersistedComponents,
 };
 use common::{
     character::{CharacterId, CharacterItem},
@@ -9,12 +9,14 @@ use common::{
 };
 use crossbeam_channel::{self, TryIter};
 use lazy_static::lazy_static;
-use std::path::Path;
-use tracing::error;
+use rusqlite::Transaction;
+use std::sync::{Arc, RwLock};
+use tracing::{error, trace};
 
-pub(crate) type CharacterListResult = Result<Vec<CharacterItem>, Error>;
-pub(crate) type CharacterCreationResult = Result<(CharacterId, Vec<CharacterItem>), Error>;
-pub(crate) type CharacterDataResult = Result<PersistedComponents, Error>;
+pub(crate) type CharacterListResult = Result<Vec<CharacterItem>, PersistenceError>;
+pub(crate) type CharacterCreationResult =
+    Result<(CharacterId, Vec<CharacterItem>), PersistenceError>;
+pub(crate) type CharacterDataResult = Result<PersistedComponents, PersistenceError>;
 type CharacterLoaderRequest = (specs::Entity, CharacterLoaderRequestKind);
 
 /// Available database operations when modifying a player's character list
@@ -53,6 +55,17 @@ pub struct CharacterLoaderResponse {
     pub result: CharacterLoaderResponseKind,
 }
 
+impl CharacterLoaderResponse {
+    pub fn is_err(&self) -> bool {
+        matches!(
+            &self.result,
+            CharacterLoaderResponseKind::CharacterData(box Err(_))
+                | CharacterLoaderResponseKind::CharacterList(Err(_))
+                | CharacterLoaderResponseKind::CharacterCreation(Err(_))
+        )
+    }
+}
+
 /// A bi-directional messaging resource for making requests to modify or load
 /// character data in a background thread.
 ///
@@ -78,84 +91,47 @@ lazy_static! {
 }
 
 impl CharacterLoader {
-    pub fn new(db_dir: &Path) -> diesel::QueryResult<Self> {
+    pub fn new(settings: Arc<RwLock<DatabaseSettings>>) -> Result<Self, PersistenceError> {
         let (update_tx, internal_rx) = crossbeam_channel::unbounded::<CharacterLoaderRequest>();
         let (internal_tx, update_rx) = crossbeam_channel::unbounded::<CharacterLoaderResponse>();
-
-        let mut conn = establish_connection(db_dir)?;
 
         let builder = std::thread::Builder::new().name("persistence_loader".into());
         builder
             .spawn(move || {
-                for request in internal_rx {
-                    let (entity, kind) = request;
+                // Unwrap here is safe as there is no code that can panic when the write lock is
+                // taken that could cause the RwLock to become poisoned.
+                let mut conn = establish_connection(&*settings.read().unwrap());
 
-                    if let Err(e) = internal_tx.send(CharacterLoaderResponse {
-                        entity,
-                        result: match kind {
-                            CharacterLoaderRequestKind::CreateCharacter {
-                                player_uuid,
-                                character_alias,
-                                persisted_components,
-                            } => CharacterLoaderResponseKind::CharacterCreation(conn.transaction(
-                                |txn| {
-                                    create_character(
-                                        &player_uuid,
-                                        &character_alias,
-                                        persisted_components,
-                                        txn,
-                                        &MATERIAL_STATS_MANIFEST,
-                                    )
-                                },
-                            )),
-                            CharacterLoaderRequestKind::DeleteCharacter {
-                                player_uuid,
-                                character_id,
-                            } => CharacterLoaderResponseKind::CharacterList(conn.transaction(
-                                |txn| {
-                                    delete_character(
-                                        &player_uuid,
-                                        character_id,
-                                        txn,
-                                        &MATERIAL_STATS_MANIFEST,
-                                    )
-                                },
-                            )),
-                            CharacterLoaderRequestKind::LoadCharacterList { player_uuid } => {
-                                CharacterLoaderResponseKind::CharacterList(conn.transaction(
-                                    |txn| {
-                                        load_character_list(
-                                            &player_uuid,
-                                            txn,
-                                            &MATERIAL_STATS_MANIFEST,
-                                        )
+                for request in internal_rx {
+                    conn.update_log_mode(&settings);
+
+                    match conn.connection.transaction() {
+                        Ok(mut transaction) => {
+                            let response =
+                                CharacterLoader::process_request(request, &mut transaction);
+                            if !response.is_err() {
+                                match transaction.commit() {
+                                    Ok(()) => {
+                                        trace!("Commit for character loader completed");
                                     },
-                                ))
-                            },
-                            CharacterLoaderRequestKind::LoadCharacterData {
-                                player_uuid,
-                                character_id,
-                            } => {
-                                let result = conn.transaction(|txn| {
-                                    load_character_data(
-                                        player_uuid,
-                                        character_id,
-                                        txn,
-                                        &MATERIAL_STATS_MANIFEST,
-                                    )
-                                });
-                                if result.is_err() {
-                                    error!(
-                                        ?result,
-                                        "Error loading character data for character_id: {}",
-                                        character_id
-                                    );
-                                }
-                                CharacterLoaderResponseKind::CharacterData(Box::new(result))
-                            },
+                                    Err(e) => error!(
+                                        "Failed to commit transaction for character loader, \
+                                         error: {:?}",
+                                        e
+                                    ),
+                                };
+                            };
+
+                            if let Err(e) = internal_tx.send(response) {
+                                error!(?e, "Could not send character loader response");
+                            }
                         },
-                    }) {
-                        error!(?e, "Could not send send persistence request");
+                        Err(e) => {
+                            error!(
+                                "Failed to start transaction for character loader, error: {:?}",
+                                e
+                            )
+                        },
                     }
                 }
             })
@@ -165,6 +141,66 @@ impl CharacterLoader {
             update_rx,
             update_tx,
         })
+    }
+
+    // TODO: Refactor the way that we send errors to the client to not require a
+    // specific Result type per CharacterLoaderResponseKind, and remove
+    // CharacterLoaderResponse::is_err()
+    fn process_request(
+        request: CharacterLoaderRequest,
+        mut transaction: &mut Transaction,
+    ) -> CharacterLoaderResponse {
+        let (entity, kind) = request;
+        CharacterLoaderResponse {
+            entity,
+            result: match kind {
+                CharacterLoaderRequestKind::CreateCharacter {
+                    player_uuid,
+                    character_alias,
+                    persisted_components,
+                } => CharacterLoaderResponseKind::CharacterCreation(create_character(
+                    &player_uuid,
+                    &character_alias,
+                    persisted_components,
+                    &mut transaction,
+                    &MATERIAL_STATS_MANIFEST,
+                )),
+                CharacterLoaderRequestKind::DeleteCharacter {
+                    player_uuid,
+                    character_id,
+                } => CharacterLoaderResponseKind::CharacterList(delete_character(
+                    &player_uuid,
+                    character_id,
+                    &mut transaction,
+                    &MATERIAL_STATS_MANIFEST,
+                )),
+                CharacterLoaderRequestKind::LoadCharacterList { player_uuid } => {
+                    CharacterLoaderResponseKind::CharacterList(load_character_list(
+                        &player_uuid,
+                        &mut transaction,
+                        &MATERIAL_STATS_MANIFEST,
+                    ))
+                },
+                CharacterLoaderRequestKind::LoadCharacterData {
+                    player_uuid,
+                    character_id,
+                } => {
+                    let result = load_character_data(
+                        player_uuid,
+                        character_id,
+                        &mut transaction,
+                        &MATERIAL_STATS_MANIFEST,
+                    );
+                    if result.is_err() {
+                        error!(
+                            ?result,
+                            "Error loading character data for character_id: {}", character_id
+                        );
+                    }
+                    CharacterLoaderResponseKind::CharacterData(Box::new(result))
+                },
+            },
+        }
     }
 
     /// Create a new character belonging to the player identified by